roachtest: add pebble write-throughput benchmark
Currently, roachtest is used to run the YCSB benchmarks on Pebble. The
results are collected and pushed to an S3 bucket for rendering on the
[Pebble Benchmarks][1] page.

A write-throughput benchmark was [recently added to Pebble][2] to aid in
identifying the maximum sustainable write throughput for a Pebble
instance and identify potential bottlenecks.

Add to the existing Pebble roachtest to register a new test similar to
the YCSB test that runs `./pebble bench write ...` on instances in the
test cluster.

The initial implementation of the write-throughput benchmark only runs a
single variant (`values=1024`) of the benchmark on GCE nodes with 16
local SSDs in a RAID 0 configuration. The test runs for eight hours. A
follow up change will add additional configurations once the new
roachtest starts consistently producing results.

Split out the common functionality from the existing
`teamcity-nightly-pebble.sh` script into a separate script that can be
sourced by the YCSB- and write-throughput-specific scripts.

Create a symlink with the exiting script name pointing to the more
specific YCSB benchmark script. This allows the existing TeamCity
configuration to continue to reference the expected file. Once TeamCity
has been updated to point at the source, this symlink can be removed.

Create a write-throughput-specific script that executes the
write-throughput benchmarks and merges the output data using `mkbench`,
similar to the YCSB script. Using separate scripts for the two benchmark
types allows the benchmark clusters to be spawned by TeamCity in
parallel, rather than serially.

Release note: None

[1]: https://cockroachdb.github.io/pebble/

[2]: cockroachdb/pebble#1345
