+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -84,7 +84,8 @@ go_library(
         "nodejs_postgres.go",
         "orm_helpers.go",
         "overload_tpcc_olap.go",
-        "pebble.go",
+        "pebble_write_throughput.go",
+        "pebble_ycsb.go",
         "pgjdbc.go",
         "pgjdbc_blocklist.go",
         "pgx.go",
