+++ pkg/cmd/roachtest/tests/registry.go
@@ -66,7 +66,8 @@ func RegisterTests(r registry.Registry) {
 	registerLibPQ(r)
 	registerLiquibase(r)
 	registerNetwork(r)
-	registerPebble(r)
+	registerPebbleWriteThroughput(r)
+	registerPebbleYCSB(r)
 	registerPgjdbc(r)
 	registerPgx(r)
 	registerNodeJSPostgres(r)
